anpr-service/requirements.txt
ultralytics==8.0.196
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

